requirements.txt
--trusted-host pypi.org 
--trusted-host files.pythonhosted.org
PyYAML==6.0
rel==0.4.7
requests==2.29.0
asyncio==3.4.3
websockets==11.0.2
fastapi==0.103.1
fastapi-websocket-rpc==0.1.25
uvicorn==0.23.2
websocket-client==1.5.1
pytest==7.3.1
pytest-order==1.1.0
pytest-tldr==0.2.5
pytest-html==3.2.0
pytest-timeout==2.1.0
pytest-repeat==0.9.1
pytest-rerunfailures==11.1.2
pytest-metadata==2.0.4
